requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
cachetools==4.2.2
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
configparser==5.0.2
crc32c==2.2.post0
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
flake8-import-order==0.18.1
fsspec==2021.5.0
future==0.18.2
gitdb==4.0.7
gitpython==3.1.17
google-auth-oauthlib==0.4.4
google-auth==1.30.0
grpcio==1.37.1
h5py==3.4.0
idna==2.10
ipykernel==5.3.4
ipython-genutils==0.2.0
ipython==7.18.1
jedi==0.17.2
joblib==1.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.3.2
markdown==3.3.4
matplotlib==3.4.3
motmetrics==1.2.0
multidict==5.1.0
numpy==1.20.3
oauthlib==3.1.0
ocrb==0.1
opencv-python==4.5.3.56
pandas==1.5.1
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.1.0
pip==21.0.1
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.17.0
psutil==5.8.0
ptyprocess==0.6.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycocotools==2.0.6
pygments==2.7.1
pyparsing==2.4.7
pytest-benchmark==3.4.1
pytest==7.2.0
python-dateutil==2.8.1
pytorch-lightning==1.1.4
pytz==2022.6
pyyaml==5.4.1
pyzmq==19.0.2
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.7.2
scikit-learn==1.1.3
scipy==1.7.1
sentry-sdk==1.1.0
setproctitle==1.3.2
setuptools==52.0.0.post20210125
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.5.0
tfrecord==1.14.1
threadpoolctl==3.1.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.0.4
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
wandb==0.13.9
watchdog==0.10.4
wcwidth==0.2.5
werkzeug==2.0.1
wheel==0.36.2
xmltodict==0.12.0
yarl==1.6.3